LNK-2723-QueryDispatch refactoring got rid of querydispatchconfiguration commands - first pass
