+++ DotNet/QueryDispatch/Application/Interfaces/IQueryDispatchConfigurationRepository.cs
@@ -3,11 +3,5 @@ using LantanaGroup.Link.Shared.Application.Repositories.Interfaces;
 
 namespace LantanaGroup.Link.QueryDispatch.Application.Interfaces
 {
-    public interface IQueryDispatchConfigurationRepository : IEntityRepository<QueryDispatchConfigurationEntity>
-    {
-        Task<QueryDispatchConfigurationEntity> GetByFacilityId(string facilityId);
-        Task<bool> DeleteByFacilityId(string facilityId);
-        Task<List<QueryDispatchConfigurationEntity>> GetAllAsync();
-        Task Update(QueryDispatchConfigurationEntity config);
-    }
+    public interface IQueryDispatchConfigurationRepository : IEntityRepository<QueryDispatchConfigurationEntity>{}
 }
