--- DotNet/QueryDispatch/Application/QueryDispatchConfiguration/Commands/ICreateQueryDispatchConfigurationCommand.cs
@@ -1,9 +1,0 @@
-﻿using LantanaGroup.Link.QueryDispatch.Domain.Entities;
-
-namespace LantanaGroup.Link.QueryDispatch.Application.QueryDispatchConfiguration.Commands
-{
-    public interface ICreateQueryDispatchConfigurationCommand
-    {
-        Task Execute(QueryDispatchConfigurationEntity config);
-    }
-}
