--- DotNet/QueryDispatch/Application/QueryDispatchConfiguration/Commands/IDeleteQueryDispatchConfigurationCommand.cs
@@ -1,9 +1,0 @@
-﻿using LantanaGroup.Link.QueryDispatch.Domain.Entities;
-
-namespace LantanaGroup.Link.QueryDispatch.Application.QueryDispatchConfiguration.Commands
-{
-    public interface IDeleteQueryDispatchConfigurationCommand
-    {
-        Task<bool> Execute(string facilityId);
-    }
-}
