--- DotNet/QueryDispatch/Application/QueryDispatchConfiguration/Commands/IUpdateQueryDispatchConfigurationCommand.cs
@@ -1,9 +1,0 @@
-﻿using LantanaGroup.Link.QueryDispatch.Domain.Entities;
-
-namespace LantanaGroup.Link.QueryDispatch.Application.QueryDispatchConfiguration.Commands
-{
-    public interface IUpdateQueryDispatchConfigurationCommand
-    {
-        Task Execute(QueryDispatchConfigurationEntity existingConfig, List<DispatchSchedule> dispatchSchedules);
-    }
-}
