--- DotNet/QueryDispatch/Application/QueryDispatchConfiguration/Queries/GetAllQueryDispatchConfigurations/IGetAllQueryDispatchConfigurationQuery.cs
@@ -1,9 +1,0 @@
-﻿using LantanaGroup.Link.QueryDispatch.Domain.Entities;
-
-namespace LantanaGroup.Link.QueryDispatch.Application.Queries
-{
-    public interface IGetAllQueryDispatchConfigurationQuery
-    {
-        Task<List<QueryDispatchConfigurationEntity>> Execute();
-    }
-}
