--- DotNet/QueryDispatch/Application/QueryDispatchConfiguration/Queries/GetQueryDispatchConfiguration/IGetQueryDispatchConfigurationQuery.cs
@@ -1,9 +1,0 @@
-﻿using LantanaGroup.Link.QueryDispatch.Domain.Entities;
-
-namespace LantanaGroup.Link.QueryDispatch.Application.Queries
-{
-    public interface IGetQueryDispatchConfigurationQuery
-    {
-        Task<QueryDispatchConfigurationEntity> Execute(string facilityId);
-    }
-}
