+++ DotNet/QueryDispatch/Presentation/Controllers/QueryDispatchController.cs
@@ -1,13 +1,14 @@
-﻿using LantanaGroup.Link.QueryDispatch.Application.Interfaces;
+﻿
+using LantanaGroup.Link.QueryDispatch.Application.Interfaces;
 using LantanaGroup.Link.QueryDispatch.Application.Models;
-using LantanaGroup.Link.QueryDispatch.Application.Queries;
-using LantanaGroup.Link.QueryDispatch.Application.QueryDispatchConfiguration.Commands;
 using LantanaGroup.Link.QueryDispatch.Domain.Entities;
 using LantanaGroup.Link.Shared.Application.Services;
 using Link.Authorization.Policies;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using QueryDispatch.Application.Settings;
+using QueryDispatch.Domain.Managers;
+using System.Threading;
 
 namespace LantanaGroup.Link.QueryDispatch.Presentation.Controllers
 {
@@ -18,21 +19,18 @@ namespace LantanaGroup.Link.QueryDispatch.Presentation.Controllers
     {
         private readonly ILogger<QueryDispatchController> _logger;
         private readonly IQueryDispatchConfigurationFactory _configurationFactory;
-        private readonly ICreateQueryDispatchConfigurationCommand _createQueryDispatchConfigurationCommand;
-        private readonly IGetQueryDispatchConfigurationQuery _getQueryDispatchConfigurationQuery;
-        private readonly IDeleteQueryDispatchConfigurationCommand _deleteQueryDispatchConfigurationCommand;
-        private readonly IUpdateQueryDispatchConfigurationCommand _updateQueryDispatchConfigurationCommand;
-        private readonly ITenantApiService _tenantApiService;
 
-        public QueryDispatchController(ILogger<QueryDispatchController> logger, IQueryDispatchConfigurationFactory configurationFactory, ICreateQueryDispatchConfigurationCommand createQueryDispatchConfigurationCommand, IGetQueryDispatchConfigurationQuery getQueryDispatchConfigurationQuery, IDeleteQueryDispatchConfigurationCommand deleteQueryDispatchConfigurationCommand, IUpdateQueryDispatchConfigurationCommand updateQueryDispatchConfigurationCommand, ITenantApiService tenantApiService)
+        private readonly ITenantApiService _tenantApiService;
+        private readonly IQueryDispatchConfigurationRepository _queryDispatchConfigRepo;
+        private readonly QueryDispatchConfigurationManager _queryDispatchConfigurationManager;
+
+        public QueryDispatchController(ILogger<QueryDispatchController> logger, IQueryDispatchConfigurationFactory configurationFactory, ITenantApiService tenantApiService, IQueryDispatchConfigurationRepository queryDispatchConfigRepo, QueryDispatchConfigurationManager queryDispatchConfigurationManager)
         {
             _logger = logger;
             _configurationFactory = configurationFactory;
-            _createQueryDispatchConfigurationCommand = createQueryDispatchConfigurationCommand;
-            _getQueryDispatchConfigurationQuery = getQueryDispatchConfigurationQuery;
-            _deleteQueryDispatchConfigurationCommand = deleteQueryDispatchConfigurationCommand;
-            _updateQueryDispatchConfigurationCommand = updateQueryDispatchConfigurationCommand;
             _tenantApiService = tenantApiService;
+            _queryDispatchConfigRepo = queryDispatchConfigRepo;
+            _queryDispatchConfigurationManager = queryDispatchConfigurationManager;
         }
 
         /// <summary>
@@ -59,7 +57,8 @@ namespace LantanaGroup.Link.QueryDispatch.Presentation.Controllers
 
             try
             {
-                var config = await _getQueryDispatchConfigurationQuery.Execute(facilityId);
+                //var config = await _getQueryDispatchConfigurationQuery.Execute(facilityId);
+                var config = await _queryDispatchConfigRepo.FirstOrDefaultAsync(x => x.FacilityId == facilityId);              
 
                 if (config == null) 
                 {
@@ -89,7 +88,7 @@ namespace LantanaGroup.Link.QueryDispatch.Presentation.Controllers
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         [HttpPost("configuration")]
-        public async Task<ActionResult<RequestResponse>> CreateQueryDispatchConfigurationAsync(QueryDispatchConfiguration model)
+        public async Task<ActionResult<RequestResponse>> CreateQueryDispatchConfigurationAsync(QueryDispatchConfiguration model, CancellationToken cancellationToken)
         {
             //validate config values
             if (model == null) 
@@ -112,7 +111,8 @@ namespace LantanaGroup.Link.QueryDispatch.Presentation.Controllers
                 }
             }
 
-            var existingConfig = await _getQueryDispatchConfigurationQuery.Execute(model.FacilityId);
+            var existingConfig = await _queryDispatchConfigRepo.FirstOrDefaultAsync(x => x.FacilityId == model.FacilityId);
+
             if (existingConfig != null)
             {
                 _logger.LogError($"Query dispatch configuration for Facility Id {model.FacilityId} was already created: {model}.");
@@ -128,7 +128,9 @@ namespace LantanaGroup.Link.QueryDispatch.Presentation.Controllers
 
                 var config = _configurationFactory.CreateQueryDispatchConfiguration(model.FacilityId, model.DispatchSchedules);
 
-                await _createQueryDispatchConfigurationCommand.Execute(config);
+
+                await _queryDispatchConfigurationManager.AddConfigEntity(config, cancellationToken);
+
 
                 return Created(config.Id, config);
             }
@@ -151,7 +153,7 @@ namespace LantanaGroup.Link.QueryDispatch.Presentation.Controllers
         [ProducesResponseType(StatusCodes.Status204NoContent)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         [HttpDelete("configuration/facility/{facilityId}")]
-        public async Task<ActionResult<RequestResponse>> DeleteQueryDispatchConfiguration(string facilityId)
+        public async Task<ActionResult<RequestResponse>> DeleteQueryDispatchConfiguration(string facilityId, CancellationToken cancellationToken)
         {
             if (string.IsNullOrEmpty(facilityId)) 
             { 
@@ -160,8 +162,8 @@ namespace LantanaGroup.Link.QueryDispatch.Presentation.Controllers
 
             try
             {
-                bool result = await _deleteQueryDispatchConfigurationCommand.Execute(facilityId);
-                
+                await _queryDispatchConfigurationManager.DeleteConfigEntity(facilityId, cancellationToken);
+
                 return NoContent();
             }
             catch (Exception ex)
@@ -188,7 +190,7 @@ namespace LantanaGroup.Link.QueryDispatch.Presentation.Controllers
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         [HttpPut("configuration/facility/{facilityId}")]
-        public async Task<ActionResult<RequestResponse>> UpdateQueryDispatchConfiguration(string facilityId, QueryDispatchConfiguration model)
+        public async Task<ActionResult<RequestResponse>> UpdateQueryDispatchConfiguration(string facilityId, QueryDispatchConfiguration model, CancellationToken cancellationToken)
         {
             if (model == null)
             {
@@ -219,18 +221,18 @@ namespace LantanaGroup.Link.QueryDispatch.Presentation.Controllers
                     return BadRequest($"Facility {facilityId} does not exist.");
                 }
 
-                var existingConfig = await _getQueryDispatchConfigurationQuery.Execute(facilityId);
+                var existingConfig = await _queryDispatchConfigRepo.FirstOrDefaultAsync(x => x.FacilityId == facilityId);
 
                 if (existingConfig == null)
                 {
                     var config = _configurationFactory.CreateQueryDispatchConfiguration(facilityId, model.DispatchSchedules);
-                    await _createQueryDispatchConfigurationCommand.Execute(config);
+                    await _queryDispatchConfigRepo.AddAsync(config);
 
                     return Created(config.Id, config);
                 }
                 else
                 {
-                    await _updateQueryDispatchConfigurationCommand.Execute(existingConfig, model.DispatchSchedules);
+                    await _queryDispatchConfigurationManager.SaveConfigEntity(existingConfig, model.DispatchSchedules, cancellationToken);
                     return NoContent();
                 }
             }
